+++ app/models/user.go
@@ -12,15 +12,15 @@ const (
 
 /**User model */
 type User struct {
-	Id       int    														// UserId 用户ID
-	UserName string `orm:"size(20)"` 						// 用户名
-	Password string `orm:"size(20)"` 						// 密码
-	Sex      int    														//性别
-	Salt     string `orm:"size(20)"`           	// 加密研制
-	Accout   string `orm:"size(20)"`           	// 账号
-	ImageUrl string `orm:"default();size(20)"` 	// 头像图片地址
-	Email    string `orm:"size(20)"`           	// 邮箱
-	Phone    string `orm:"size(20)"`           	// 用户电话
-	Address  string `orm:"size(20)"`           	// 用户地址
-	Status   int    														// 用户状态
+	Id       int64  // UserId 用户ID
+	UserName string `orm:"size(20)"` // 用户名
+	Password string `orm:"size(20)"` // 密码
+	Sex      int    //性别
+	Salt     string `orm:"size(20)"`           // 加密研制
+	Accout   string `orm:"size(20)"`           // 账号
+	ImageUrl string `orm:"default();size(20)"` // 头像图片地址
+	Email    string `orm:"size(20)"`           // 邮箱
+	Phone    string `orm:"size(20)"`           // 用户电话
+	Address  string `orm:"size(20)"`           // 用户地址
+	Status   int    // 用户状态
 }
